+++ NINA.Plugin.TargetScheduler/Database/Migrate/21.sql
@@ -1,0 +1,6 @@
+/*
+*/
+
+ALTER TABLE profilepreference ADD COLUMN enableStopOnHumidity INTEGER DEFAULT 1;
+
+PRAGMA user_version = 21;